requirements.txt
aiocache==0.12.3
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.8
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anthropic==0.43.0
anyio==4.8.0
appdirs==1.4.4
APScheduler==3.10.4
argcomplete==3.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
Authlib==1.3.2
av==12.3.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
beniget==0.4.1
bidict==0.23.1
binwalk==2.3.3
bitarray==2.9.3
bitstring==4.2.3
black==24.8.0
blinker==1.9.0
boto3==1.35.53
botocore==1.35.99
Brlapi==0.8.4
Brotli==1.0.9
build==1.2.2.post1
cachetools==5.5.0
cattrs==24.1.2
certifi==2024.12.14
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.0.1
chroma-hnswlib==0.7.6
chromadb==0.5.15
cliapp==1.20180812.1
click==8.1.8
cloup==3.0.5
cmake-language-server==0.1.10
cmdtest==0.32+git
colbert-ai==0.2.21
colorama==0.4.6
colorclass==2.2.2
coloredlogs==15.0.1
compressed-rtf==1.0.6
contourpy==1.0.7
coverxygen==1.8.1
crit==3.17.1
cryptography==44.0.0
cssselect==1.2.0
ctranslate2==4.5.0
cupshelpers==1.0
cycler==0.11.0
dataclasses-json==0.6.7
datasets==3.2.0
dbus-python==1.3.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
dill==0.3.8
distro==1.8.0
dnspython==2.7.0
docker==7.1.0
docx2txt==0.8
duckduckgo_search==6.3.7
durationpy==0.9
easygui==0.98.3
ebcdic==1.1.1
ecdsa==0.19.0
einops==0.8.0
email_validator==2.2.0
emoji==2.14.1
esptool==4.8.1
et_xmlfile==2.0.0
eval_type_backport==0.2.2
evdev==1.6.1
Events==0.5
extract-msg==0.52.0
fake-useragent==1.5.1
fastapi==0.111.0
fastapi-cli==0.0.7
faster-whisper==1.0.3
filelock==3.16.1
filetype==1.2.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==24.12.23
fonttools==4.55.3
fpdf2==2.8.2
frozenlist==1.5.0
fs==2.4.16
fsspec==2024.9.0
ftfy==6.2.3
fuse-python==1.0.5
gast==0.5.2
git-python==1.0.3
gitdb==4.0.12
GitPython==3.1.44
glcontext==3.0.0
google-ai-generativelanguage==0.6.6
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
gpg==1.18.0
greenlet==2.0.2
grpcio==1.67.1
grpcio-status==1.62.3
grpcio-tools==1.62.3
gyp==0.1
h11==0.14.0
h2==4.1.0
hidapi==0.9.0.post3
hpack==4.0.0
html5lib==1.1
httpcore==1.0.7
httpie==3.2.1
httplib2==0.20.4
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
imageio==2.36.1
imageio-ffmpeg==0.5.1
importlib_metadata==8.4.0
importlib_resources==6.5.2
iniconfig==2.0.0
intelhex==2.3.0
isosurfaces==0.1.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kiwisolver==0.0.0
kubernetes==31.0.0
langchain==0.3.7
langchain-chroma==0.1.4
langchain-community==0.3.7
langchain-core==0.3.30
langchain-text-splitters==0.3.5
langdetect==1.0.9
langfuse==2.44.0
langsmith==0.1.147
lark==1.1.9
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
ldap3==2.9.1
libevdev==0.5
libvirt-python==9.0.0
louis==3.24.0
lsprotocol==2023.0.1
lxml==5.3.0
lz4==4.0.2+dfsg
Mako==1.3.8
manim==0.18.1
ManimPango==0.6.0
mapbox_earcut==1.0.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib==3.6.3
mdurl==0.1.2
meson==1.7.2
milvus-lite==2.4.11
mmh3==5.0.1
moderngl==5.12.0
moderngl-window==3.0.3
monotonic==1.6
moviepy==2.1.1
mpmath==1.3.0
msgpack==1.0.3
msoffcrypto-tool==5.4.2
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
olefile==0.47
oletools==0.60.2
ollama==0.4.7
onnxruntime==1.20.1
onnxruntime-gpu==1.20.1
open-webui==0.5.4
openai==1.59.7
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
opensearch-py==2.7.1
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.14
overrides==7.7.0
packaging==23.2
pandas==2.2.3
passlib==1.7.4
pathspec==0.12.1
pcodedmp==1.2.6
peewee==3.17.8
peewee-migrate==1.12.2
pgvector==0.3.5
pillow==10.4.0
platformdirs==4.3.6
plover==4.0.0.dev10
pluggy==1.5.0
ply==3.11
portalocker==2.10.1
posthog==3.8.3
primp==0.10.0
proglog==0.1.10
propcache==0.2.1
proto-plus==1.25.0
protobuf==4.25.5
protontricks==1.10.2
psutil==6.1.1
psycopg2-binary==2.9.9
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycairo==1.20.1
pyclipper==1.3.0.post6
pycparser==2.22
pycups==2.0.1
pycurl==7.45.2
pydantic==2.9.2
pydantic-settings==2.7.1
pydantic_core==2.23.4
pydub==0.25.1
pyglet==2.0.20
PyGLM==2.7.3
pygls==1.3.1
Pygments==2.14.0
PyGObject==3.42.2
PyJWT==2.10.1
pylibacl==0.7.0
pymdown-extensions==10.11.2
pymilvus==2.5.0
pymongo==4.10.1
PyMySQL==1.1.1
pynvim==0.4.2
PyOpenGL==3.1.6
pypandoc==1.13
pyparsing==3.0.9
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PyQt5==5.15.9
PyQt5-sip==12.11.1
PyQt6==6.4.2
PyQt6-sip==13.4.1
pyqtgraph==0.13.1
pyserial==3.5
PySimpleSOAP==1.16.2
pysmbc==1.0.23
PySocks==1.7.1
pyte==0.8.0
pytest==8.3.4
pytest-docker==3.1.1
python-apt==2.6.0
python-dateutil==2.9.0.post0
python-debian==0.1.49
python-debianbts==4.0.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-iso639==2024.10.22
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.18
python-oxmsg==0.0.1
python-pptx==1.0.0
python-socketio==5.11.3
python-xlib==0.33
pythran==0.11.0
pytube==15.0.0
pytz==2024.2
pyudev==0.24.0
pyxattr==0.8.1
pyxdg==0.28
pyxlsb==1.0.10
PyYAML==6.0.2
qdrant-client==1.12.2
rank-bm25==0.2.2
RapidFuzz==3.11.0
rapidocr-onnxruntime==1.3.24
red-black-tree-mod==1.20
redis==5.2.1
reedsolo==1.7.0
regex==2024.11.6
reportbug==12.0.0
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
RTFDE==0.1.2
s3transfer==0.10.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.0
scour==0.38.2
screeninfo==0.8.1
sentence-transformers==3.3.1
sentencepiece==0.2.0
setproctitle==1.3.1
shapely==2.0.6
shellingham==1.5.4
simple-websocket==1.1.0
six==1.16.0
skia-pathops==0.8.0.post2
smmap==5.0.2
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
SQLAlchemy==2.0.32
srt==3.5.3
starlette==0.37.2
svgelements==1.9.6
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tornado==6.2
tqdm==4.67.1
transformers==4.48.0
triton==3.1.0
ttystatus==0.38
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ufoLib2==0.14.0
ufw==0.36.2
ujson==5.10.0
unstructured==0.15.9
unstructured-client==0.28.1
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.30.6
uvloop==0.21.0
validators==0.34.0
vdf==3.4
vulkan-object==1.4.323
wadllib==1.3.6
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
wxPython==4.2.0
xdg==5
xlrd==2.0.1
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.18.3
youtube-transcript-api==0.6.3
zipp==3.21.0
